backend/pyproject.toml
[tool.poetry]
name = "speech-slurring-detection"
version = "1.0.0"
description = "Speech Slurring / Dysarthria Detection System for Brain Stroke Early Detection"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# ── Core ML ────────────────────────────────────────────────────────────────
torch = "^2.6.0"
torchaudio = "^2.6.0"
transformers = "^5.0.0"

# ── Audio Processing ────────────────────────────────────────────────────────
librosa = "^0.11.0"
praat-parselmouth = "^0.4.7"
torchcrepe = "^0.0.24"
opensmile = "^2.6.0"
silero-vad = "^6.1.0"
pyloudnorm = "^0.1.1"
scipy = "^1.11.0"
soundfile = "^0.12.0"

# ── API ─────────────────────────────────────────────────────────────────────
fastapi = "^0.115.0"
pydantic = "^2.8.0"
pydantic-settings = "^2.0.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-multipart = "^0.0.9"
aiofiles = "^24.0.0"

# ── Reporting ───────────────────────────────────────────────────────────────
reportlab = "^4.2.0"
weasyprint = "^62.0"
jinja2 = "^3.1.0"
matplotlib = "^3.9.0"
plotly = "^6.0.0"
kaleido = "^0.2.1"
pillow = "^11.0.0"

# ── MLOps ───────────────────────────────────────────────────────────────────
mlflow = "^3.0.0"
evidently = "^0.7.0"

# ── Utilities ───────────────────────────────────────────────────────────────
numpy = "^1.26.0"
pandas = "^2.2.0"
pyyaml = "^6.0.0"
python-dotenv = "^1.0.0"
redis = "^5.0.0"
httpx = "^0.27.0"

# ── Monitoring ──────────────────────────────────────────────────────────────
prometheus-client = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.27.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
ipython = "^8.26.0"
ipykernel = "^6.29.0"
jupyterlab = "^4.2.0"
black = "^24.8.0"

[tool.poetry.scripts]
train-hubert = "training.train_hubert_salr:main"
train-cnn = "training.train_cnn_bilstm:main"
evaluate = "training.evaluate:main"
serve = "api.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py311']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
